refactor:разграничить ответственность между сервисами и хранилищами
