+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -5,9 +5,12 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.AlreadyExistException;
 import ru.yandex.practicum.filmorate.exception.IncorrectParameterException;
 import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,6 +21,34 @@ public class UserService {
 
     public UserService(UserStorage userStorage) {
         this.userStorage = userStorage;
+    }
+
+    public User addUser(User user) throws Exception {
+        if (!validation(user)) {
+            log.info("Ошибка валидации");
+            throw new ValidationException("Ошибка валидации");
+        }
+        if (userStorage.getUsers().containsValue(user)) {
+            log.info("Добавление через POST-запрос уже имеющегося объекта");
+            throw new AlreadyExistException("Данный пользователь уже добавлен");
+        }
+        return userStorage.addUser(user);
+    }
+
+    public User addOrUpdateUser(User user) throws Exception {
+        if (!validation(user)) {
+            log.info("Ошибка валидации");
+            throw new ValidationException("Ошибка валидации");
+        }
+        User testUser = userStorage.addOrUpdateUser(user);
+        if (testUser == null) {
+            throw new UserNotFoundException("Попытка обновления предварительно не добавленного объекта");
+        }
+        return testUser;
+    }
+
+    public List<User> getAll() {
+        return new ArrayList<>(userStorage.getUsers().values());
     }
 
     public void addFriend(Integer userId, Integer requestedUserId) {
@@ -57,7 +88,7 @@ public class UserService {
             throw new UserNotFoundException("Пользователь с данным id не найден");
         }
         List<User> userFriends = new ArrayList<>();
-        for (User userInStorage : userStorage.getAll()) {
+        for (User userInStorage : getAll()) {
             for (Integer id : userStorage.getUsers().get(userId).getFriendsId()) {
                 if (userInStorage.getId().equals(id)) {
                     userFriends.add(userStorage.getUsers().get(id));
@@ -90,5 +121,17 @@ public class UserService {
             log.info("Пользователь с данным id не найден");
             throw new UserNotFoundException("Пользователь с данным id не найден");
         }
+    }
+
+    private boolean validation(User user) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        return user.getEmail() != null
+                && !user.getEmail().isBlank()
+                && !user.getEmail().isEmpty()
+                && user.getEmail().contains("@")
+                && !user.getLogin().isBlank()
+                && !user.getLogin().isEmpty()
+                && !user.getLogin().contains(" ")
+                && !LocalDate.parse(user.getBirthday(), formatter).isAfter(LocalDate.now());
     }
 }
