+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.storage.film;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
-import java.util.List;
 import java.util.Map;
 
 public interface FilmStorage {
@@ -11,6 +10,4 @@ public interface FilmStorage {
     Film addFilm(Film film) throws Exception;
 
     Film addOrUpdateFilm(Film film) throws Exception;
-
-    List<Film> getAll();
 }
