+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -3,17 +3,9 @@ package ru.yandex.practicum.filmorate.storage.user;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.RequestBody;
-import ru.yandex.practicum.filmorate.exception.AlreadyExistException;
-import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @Slf4j
@@ -23,15 +15,7 @@ public class InMemoryUserStorage implements UserStorage {
     private final Map<Integer, User> users = new HashMap<>();
     private int id = 0;
 
-    public User addUser(@RequestBody User user) throws Exception {
-        if (!validation(user)) {
-            log.info("Ошибка валидации");
-            throw new ValidationException("Ошибка валидации");
-        }
-        if (users.containsValue(user)) {
-            log.info("Добавление через POST-запрос уже имеющегося объекта");
-            throw new AlreadyExistException("Данный пользователь уже добавлен");
-        }
+    public User addUser(User user) {
         if (user.getName() == null || user.getName().isEmpty() || user.getName().isBlank()) {
             user.setName(user.getLogin());
         }
@@ -42,11 +26,7 @@ public class InMemoryUserStorage implements UserStorage {
         return user;
     }
 
-    public User addOrUpdateUser(@RequestBody User user) throws Exception {
-        if (!validation(user)) {
-            log.info("Ошибка валидации");
-            throw new ValidationException("Ошибка валидации");
-        }
+    public User addOrUpdateUser(User user) {
         if (user.getName().isEmpty() || user.getName().isBlank()) {
             user.setName(user.getLogin());
         }
@@ -62,24 +42,8 @@ public class InMemoryUserStorage implements UserStorage {
                 users.put(user.getId(), user);
                 return user;
             } else {
-                throw new UserNotFoundException("Попытка обновления предварительно не добавленного объекта");
+                return null;
             }
         }
-    }
-
-    public List<User> getAll() {
-        return new ArrayList<>(users.values());
-    }
-
-    private boolean validation(User user) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        return user.getEmail() != null
-                && !user.getEmail().isBlank()
-                && !user.getEmail().isEmpty()
-                && user.getEmail().contains("@")
-                && !user.getLogin().isBlank()
-                && !user.getLogin().isEmpty()
-                && !user.getLogin().contains(" ")
-                && !LocalDate.parse(user.getBirthday(), formatter).isAfter(LocalDate.now());
     }
 }
