+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.util.List;
 import java.util.Map;
 
 public interface UserStorage {
@@ -11,6 +10,4 @@ public interface UserStorage {
     User addUser(User user) throws Exception;
 
     User addOrUpdateUser(User user) throws Exception;
-
-    List<User> getAll();
 }
